seperate AddTask, TaskItem, TaskList from App
