+++ frontend/src/components/TaskList.tsx
@@ -1,0 +1,34 @@
+import React, { useState, useEffect } from "react";
+import { Box, List } from "@chakra-ui/react";
+import axios from "axios";
+import TaskItem from "./TaskItem";
+
+interface Task {
+  id: number;
+  title: string;
+  completed: boolean;
+}
+
+const TaskList: React.FC = () => {
+  const [tasks, setTasks] = useState<Task[]>([]);
+
+  useEffect(() => {
+    const fetchTasks = async () => {
+      const response = await axios.get("http://localhost:3005/tasks");
+      setTasks(response.data);
+    };
+    fetchTasks();
+  }, []);
+
+  return (
+    <Box m={10}>
+      <List spacing={3} mt="4">
+        {tasks.map((task) => (
+          <TaskItem key={task.id} task={task} />
+        ))}
+      </List>
+    </Box>
+  );
+};
+
+export default TaskList;
